# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Snipr URL shortener API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /links:
    get:
      operationId: getLinks
      tags: [links]
      summary: Get all links for current workspace
      parameters:
        - name: folderId
          in: query
          schema:
            type: string
        - name: tagId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Link"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createLink
      tags: [links]
      summary: Create a new short link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLinkRequest"
      responses:
        "201":
          description: Link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "409":
          description: Slug already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /links/{id}:
    get:
      operationId: getLink
      tags: [links]
      summary: Get a link by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Link found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateLink
      tags: [links]
      summary: Update a link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLinkRequest"
      responses:
        "200":
          description: Link updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteLink
      tags: [links]
      summary: Delete a link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /links/{id}/qr:
    get:
      operationId: getLinkQr
      tags: [links]
      summary: Get QR code SVG for a link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: QR code data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Link:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        slug:
          type: string
        destinationUrl:
          type: string
        title:
          type: string
          nullable: true
        enabled:
          type: boolean
        expiresAt:
          type: string
          format: date-time
          nullable: true
        folderId:
          type: string
          nullable: true
        clickLimit:
          type: number
          nullable: true
        fallbackUrl:
          type: string
          nullable: true
        hasPassword:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - workspaceId
        - slug
        - destinationUrl
        - enabled
        - hasPassword
        - createdAt
        - updatedAt
    CreateLinkRequest:
      type: object
      properties:
        slug:
          type: string
          nullable: true
          description: "Custom slug. If null, auto-generated."
        destinationUrl:
          type: string
          format: uri
        title:
          type: string
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        folderId:
          type: string
          nullable: true
        clickLimit:
          type: number
          nullable: true
        fallbackUrl:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
      required:
        - destinationUrl
    UpdateLinkRequest:
      type: object
      properties:
        slug:
          type: string
          nullable: true
        destinationUrl:
          type: string
          format: uri
          nullable: true
        title:
          type: string
          nullable: true
        enabled:
          type: boolean
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        folderId:
          type: string
          nullable: true
        clickLimit:
          type: number
          nullable: true
        fallbackUrl:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
          description: "Set to empty string to remove password"
    QrResponse:
      type: object
      properties:
        svg:
          type: string
        shortUrl:
          type: string
      required:
        - svg
        - shortUrl
...